moved some handler to monolog bridge
