+++ src/Viserio/Bridge/Monolog/.gitattributes
@@ -1,0 +1,4 @@
+Tests/           export-ignore
+phpunit.xml.dist export-ignore
+.gitattributes   export-ignore
+.gitignore       export-ignore
