+++ src/Viserio/Bridge/Monolog/.gitignore
@@ -1,0 +1,3 @@
+vendor/
+composer.lock
+phpunit.xml
