+++ src/Viserio/Bridge/Monolog/README.md
@@ -1,4 +1,4 @@
-<h1 align="center">Viserio Foundation Component</h1>
+<h1 align="center">Viserio Monolog Bridge</h1>
 <p align="center">
     <a href="https://github.com/narrowspark/framework/releases"><img src="https://img.shields.io/packagist/v/narrowspark/framework.svg?style=flat-square"></a>
     <a href="https://php.net/"><img src="https://img.shields.io/badge/php-%5E7.2.0-8892BF.svg?style=flat-square"></a>
@@ -17,7 +17,7 @@ Installation
 Use [Composer](https://getcomposer.org/) to install this package:
 
 ```sh
-composer require viserio/foundation
+composer require viserio/monolog-bridge
 ```
 
 Official Documentation
