+++ src/Viserio/Component/Log/composer.json
@@ -28,17 +28,15 @@
         "mockery/mockery"                      : "^1.0",
         "narrowspark/testing-helper"           : "^6.0",
         "phpunit/phpunit"                      : "^6.0",
-        "symfony/var-dumper"                   : "~4.0",
-        "viserio/console"                      : "self.version",
         "viserio/container"                    : "self.version",
         "viserio/events"                       : "self.version",
         "viserio/options-resolver"             : "self.version"
     },
     "autoload": {
         "psr-4": {
-            "Viserio\\Component\\Log\\"       : ""
+            "Viserio\\Component\\Log\\"        : ""
         },
-        "exclude-from-classmap"               : ["/Tests/"]
+        "exclude-from-classmap"                : ["/Tests/"]
     },
     "autoload-dev": {
         "psr-4": {
@@ -50,9 +48,7 @@
     },
     "suggest": {
         "container-interop/service-provider"   : "Required to use service-provider (^0.4).",
-        "symfony/var-dumper"                   : "Required to use VarDumperFormatter and optional for ConsoleFormatter (~3.4).",
-        "viserio/console"                      : "Required to use ConsoleFormatter (^self.version).",
-        "viserio/events"                       : "Required to use ConsoleFormatter (^self.version).",
+        "viserio/events"                       : "Required for log messages in logger (^self.version).",
         "viserio/statical-proxy"               : "Required to use the proxy instances (self.version).",
         "viserio/options-resolver"             : "Required to use service-provider (self.version)."
     },
